Make `GenerateNonce` and `GenerateNonceVerifier` public
