+++ Gotrue/Exceptions/RequestException.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Net.Http;
+using Supabase.Gotrue.Responses;
+
+namespace Supabase.Gotrue
+{
+    public class RequestException : Exception
+    {
+        public HttpResponseMessage Response { get; private set; }
+        public ErrorResponse Error { get; private set; }
+
+        public RequestException(HttpResponseMessage response, ErrorResponse error) : base(error.Message)
+        {
+            Response = response;
+            Error = error;
+        }
+    }
+}
